# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Financial Advisor - Web Search Service",
    "description": "A small FastAPI microservice that provides a unified web search API for the AI Financial Advisor agent. This service should be used for general information lookup that is *not* strictly tied to the Trade Engine's quantitative market data.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/dataapi-query": {
      "get": {
        "tags": ["admin"],
        "summary": "Dataapi Query",
        "description": "Proxy a read-only SQL query to the DataAPI admin endpoint.",
        "operationId": "dataapi_query_api_admin_dataapi_query_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 2000,
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {},
                  "type": "object",
                  "title": "Response Dataapi Query Api Admin Dataapi Query Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                { "type": "string" },
                { "type": "integer" }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}